Changed how BestHandIdentifier does its work.
Instead of doing all of the heavy lifting of figuring out which of
all the possible hands the cards could best become, it now delegates
to the various hand specifications.  It'll just do a search down the
list of specifications and the first match it finds will be the best
hand.  So, the order of the specifications' evaluation is important.
That sequence is a lot easier to express and see than the individual
bundles of logic they all represent.

For now, BHI news up all of its specifications, the class could/
should be further changed to take them all as dependencies.