Changed FullHouseSpecification's mission somewhat.
Instead of just handling whether a Hand can be a FullHouse, it'll
take a Hand whose cards could be a FullHouse and return a FullHouse
without any other cards.  If such a hand could not be made, it instead
returns null.  To stay compatible with its old function, satisfying the
specification is a simple matter of running Hands through the newHand
method and testing what comes out the other end.