StraightSpecification receives newHand().
StraightSpecification should now be able to return the best
Straight possible given a Hand with any number of Cards.

The isSatisfiedBy(Hand) method was able to remain unchanged, but I
had to override _canBeAStraight() here in order to hook newHand()'s
behavior into into it.  I did it this way since _canBeAStraight()
is also used elsewhere.