StraightSpecification::newHand needed some more work.
Once the cards are sorted, they're guaranteed to be in order.
That being the case, if the next card isn't one less, the cards
accumulated so far cannot form a Straight.

... THAT IS, unless, the next card has the same face value as the
previous card.  In which case, we can discard it, and keep looking
for the one less in the next card.