Removed two tests that don't seem to be relevant anymore.
In order for specs to be composable, they need to only know about
what cards can form a valid hand of the type for which they were
written - take the StraightSpecification - should it be satisfied
by StraightFlush or RoyalFlush?  I'm disagreeing with my past
decisions when now I say yes, it should.  Knowledge about the BEST
hand to play should belong to BHI, not to individual specs.