Relaxed StraightSpecification's requirements.
Since I'm removing the responsibility of making sure the Hand is
the best possible Hand it could be before it will satisfy the spec,
I can just reduce the requirements - StraightSpecification is
satisfied by any hand that can be a Straight.  Who cares if it
can also be a flush? Not StraightSpecification, at least. He hopes
that there's a FlushSpecification out there that can worry about that.
Maybe there's even some other service out there, perhaps a
BestHandIdentifier, that can worry about which specification needs
to be satisfied first in order for the best hand to be identified....