Fixed StraightSpecification wrt Wheel.
The new way to validate Hands by trying to build them from the cards you
were given broke with regard to Straight.

The algorithm for Straight just needed a slight tweak - assuming a
sorted array of cards, if the previous card was an ace and the next card
is a 5, that could possibly be a Wheel, so let that slide.