Convert FlushSpecificationTest from php to C#.
I regret not being able to replicate the crazy array generation I was
performing in getSomePossibleFlushes(), but right this moment, I don't
think any extra effort on it was warranted.  For now, a hard-coded list
of a few Flushes will suffice.  If time permits, I will revisit this at
some point in the future.