Get BestHandIdentifierTest to Compile, Part III.
I guess Microsoft discourages this type of reflection, or at least
recognizes the potential security risks it presents, since when all you
have is a string of the classname you want to instantiate, and not just
call the default constructor (Hand has no such constructor,
incidentally), then you have to use this big bad overloaded method to do
it.  The MSDN docs don't mention it explicitly, but stackoverflow and my
own experiences confirm that you can pass null for all the other
parameters and apparently the proper defaults will be used.