Return to CardBuilderTest's original intent.
I wasn't able to get any kind of equality assertion to work on the
expected and actual card objects, regardless of the fact that they were
in fact the same card values, as far as their face value and suit were
concerned.  That being the case, why try to compare anything else besides
their ToString() contents?