Deck needs to know when it already contains a card.
It seems that the card objects are being tested for equality by
comparing their actual references, i.e. that the two objects are in fact
the same reference.  That's not what I want at all.  I consider two
cards to be equal if they both evaluate to the same string.