Rename a variable to have a name other than its type.
Having values the same name as their type is a bad idea.
Another one of my php holdovers, a habit I need to break in other
languages (yes, I know, this could be somewhat excused as an artifact of
the transliteration process from php to C#) is to give different
identifiers to my objects than what their types are declared to be.